+++ chats/apps/api/v1/routers.py
@@ -10,9 +10,14 @@ from chats.apps.api.v1.external.rooms.viewsets import (
     CustomFieldsUserExternalViewSet,
     ExternalListRoomsViewSet,
     RoomFlowViewSet,
+    RoomMetricsViewSet,
     RoomUserExternalViewSet,
 )
 from chats.apps.api.v1.external.sectors.viewsets import SectorFlowViewset
+from chats.apps.api.v1.groups_sectors.viewsets import (
+    GroupSectorAuthorizationViewset,
+    GroupSectorViewset,
+)
 from chats.apps.api.v1.internal.dashboard.viewsets import InternalDashboardViewset
 # flake8: noqa: F811
 from chats.apps.api.v1.internal.projects import viewsets as project_internal_views
@@ -25,6 +30,8 @@ from chats.apps.api.v1.projects.viewsets import (
     CustomStatusTypeViewSet,
     CustomStatusViewSet,
 )
+from chats.apps.api.v1.orgs.viewsets import OrgProjectViewSet
+from chats.apps.api.v1.projects.viewsets import ProjectPermissionViewset, ProjectViewset
 from chats.apps.api.v1.queues.viewsets import QueueAuthorizationViewset, QueueViewset
 from chats.apps.api.v1.quickmessages.viewsets import (
     QuickMessageViewset,
@@ -109,7 +116,17 @@ router.register("contact", ContactViewset)
 router.register("history/rooms", HistoryRoomViewset, basename="history_room")
 router.register("discussion", DiscussionViewSet, basename="discussion")
 router.register("sector", SectorViewset)
+router.register("group_sector", GroupSectorViewset, basename="group_sector")
+router.register(
+    "authorization/group_sector",
+    GroupSectorAuthorizationViewset,
+    basename="group_sector_auth",
+)
 router.register("tag", SectorTagsViewset)
+router.register(
+    r"org/(?P<org>[\w-]+)/projects", OrgProjectViewSet, basename="org-projects"
+)
+
 router.register("project", ProjectViewset)
 router.register("custom_status", CustomStatusViewSet)
 router.register("custom_status_type", CustomStatusTypeViewSet)
@@ -162,6 +179,11 @@ router.register(
     "external/custom_field",
     CustomFieldsUserExternalViewSet,
     basename="external_custom_field",
+)
+router.register(
+    "external/rooms_metrics",
+    RoomMetricsViewSet,
+    basename="external_room_metrics",
 )
 router.register(
     "external/list_rooms",