+++ chats/apps/rooms/tests/test_models.py
@@ -1,4 +1,5 @@
 import uuid
+
 from django.db import IntegrityError
 from django.test import TestCase
 from django.utils import timezone