+++ chats/urls.py
@@ -20,16 +20,19 @@ from django.conf import settings
 from django.conf.urls import include
 from django.conf.urls.static import static
 from django.contrib import admin
+from django.http import HttpResponse
 from django.urls import path, re_path
 from django.views.static import serve
+
 from chats.apps.api.v1.swagger import schema_view
-from django.http import HttpResponse
+from chats.apps.api.v1.prometheus.views import metrics_view
 
 urlpatterns = [
     path("", lambda _: HttpResponse()),
     path("doc/", schema_view.with_ui("redoc", cache_timeout=0), name="schema-redoc"),
     path("admin/", admin.site.urls),
     path("v1/", include("chats.apps.api.v1.urls")),
+    path("api/prometheus/metrics", metrics_view, name="metrics_view"),
 ]
 
 # Static files