+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# 3.16.0
+## Add
+  - Add group sector
+ 
+# 3.15.3
+## Add
+  - Add new contact urn to rooms metrics
+
+# 3.15.2
+## Add
+  - New endpoint to list rooms metrics
+
+# 3.15.1
+## Fix
+  - Export chats dashboard data to CSV and XLS
+
+# 3.15.0
+## Add
+  - New endpoint to set project as primary project to be used in infracommerce integration.
+## Add
+  - Endpoint to integrate primary and secondary projects
+## Add
+   - Adding prometheus endpoint for monitoring the application.
+
 # 3.14.0
 ## Add
   - Introduced a new publisher that triggers when a room is closed